% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{build_date}
\alias{build_date}
\title{Build Date Strings and Retrieve Manuscript Dates}
\usage{
build_date(date_node)
}
\arguments{
\item{date_node}{An XML node representing a date element.}

\item{pubmed_article_element}{An XML element representing a PubMed article.}
}
\value{
A character string containing the formatted date or "N/A" if not available.
}
\description{
These internal functions build date strings and retrieve manuscript dates (received and accepted) from PubMed articles.
They are used to process the XML elements returned by the PubMed API.
}
\keyword{internal}
