% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlink.R
\name{fetch_pubmed_outlinks}
\alias{fetch_pubmed_outlinks}
\title{Fetches outbound links for PubMed IDs}
\usage{
fetch_pubmed_outlinks(db, ids)
}
\arguments{
\item{db}{A character string indicating the database to fetch the outbound links from.
This should be "pubmed" for PubMed.}

\item{ids}{A numeric or character vector containing the PubMed IDs for which outbound links need to be fetched.}
}
\value{
An XML document containing the outbound links for the given PubMed IDs.
}
\description{
This function fetches the outbound links from PubMed database for the given PubMed IDs. It constructs a URL using the PubMed API and fetches the XML response containing the outbound links.
}
\examples{
\dontrun{
outlinks <- fetch_pubmed_outlinks("pubmed", c(12345678, 23456789))
}
}
\keyword{internal}
