% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_funcs.R
\name{find_pubmed_ids}
\alias{find_pubmed_ids}
\title{Find PubMed Article IDs}
\usage{
find_pubmed_ids(db = "pubmed", query, retmax, retstart)
}
\arguments{
\item{db}{A character string specifying the database to search in. Default is "pubmed".}

\item{query}{A character string specifying the search query for finding article IDs.}

\item{retmax}{An integer specifying the maximum number of article IDs to return. Default is 100.}

\item{retstart}{An integer specifying the starting position for the search. Default is 0.}
}
\value{
A character vector containing the PubMed article IDs that match the search query.
}
\description{
This internal function searches for PubMed article IDs based on the given query.
}
\keyword{internal}
