% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{get_content}
\alias{get_content}
\title{Retrieve Content from the PubMed Database}
\usage{
get_content(db = "pubmed", query, max_rows)
}
\arguments{
\item{db}{A character string specifying the database to search in. Default is "pubmed".}

\item{query}{A character string specifying the search query for finding articles.}

\item{max_rows}{An integer specifying the maximum number of articles to return.}
}
\value{
A data frame containing content from the specified PubMed articles.
}
\description{
This function retrieves content from the PubMed database based on a given search query and
the maximum number of rows to be returned.
}
\examples{
\dontrun{
# Retrieve content from PubMed articles related to "cancer" (max 100 articles)
cancer_articles <- get_content(db = "pubmed", query = "cancer", max_rows = 100)
}

}
