% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_funcs.R
\name{get_key_opt}
\alias{get_key_opt}
\title{Utility function to retrieve the NCBI API key from R options}
\usage{
get_key_opt()
}
\description{
Utility function to retrieve the NCBI API key from R options
}
