% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_titles.R
\name{get_xml_text_or_default}
\alias{get_xml_text_or_default}
\alias{get_pubmed_article_title}
\alias{get_pubmed_article_journal}
\title{Retrieve XML Text or Default Value}
\usage{
get_xml_text_or_default(xml_element, default)

get_pubmed_article_title(pubmed_article_element)

get_pubmed_article_journal(pubmed_article_element)
}
\arguments{
\item{xml_element}{An XML element from which to retrieve text.}

\item{default}{A default value to return if the XML element is not available.}

\item{pubmed_article_element}{An XML element representing a PubMed article.}
}
\value{
A character string containing the text from the XML element or the default value.

A character string containing the article or journal title or "N/A" if not available.
}
\description{
These internal functions retrieve text from XML elements or return a default value if the element is not available.
They are used to process the XML elements returned by the PubMed API.

These internal functions retrieve article and journal titles from PubMed articles.
They are used to process the XML elements returned by the PubMed API.
}
\keyword{internal}
