% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{get_authors}
\alias{get_authors}
\alias{get_country}
\alias{doi_results}
\alias{get_pmid}
\title{Retrieve Information from PubMed Articles}
\usage{
get_authors(pubmed_article_element)

get_country(pubmed_article_element)

doi_results(pubmed_article_element)

get_pmid(pubmed_article_element)
}
\arguments{
\item{pubmed_article_element}{An XML element representing a PubMed article.}
}
\value{
A character string containing the requested information or "N/A" if not available.
}
\description{
These internal functions retrieve various information such as authors, country of publication, DOIs, and PMIDs from PubMed articles.
They are used to process the XML elements returned by the PubMed API.
}
\keyword{internal}
