% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_articles.R
\name{process_pubmed_article}
\alias{process_pubmed_article}
\title{Process a PubMed Article Element}
\usage{
process_pubmed_article(pubmed_article_element)
}
\arguments{
\item{pubmed_article_element}{An XML element representing a PubMed article.}
}
\value{
A data frame containing information about the processed PubMed article.
}
\description{
This internal function processes a PubMed article element and extracts relevant information.
}
\keyword{internal}
