% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_articles.R
\name{process_pubmed_articleset}
\alias{process_pubmed_articleset}
\title{Process a Set of PubMed Articles}
\usage{
process_pubmed_articleset(pubmed_article_set)
}
\arguments{
\item{pubmed_article_set}{An XML document containing a set of PubMed articles.}
}
\value{
A data frame containing information about the processed PubMed articles.
}
\description{
This internal function processes a set of PubMed articles and extracts relevant information from the XML elements.
}
\keyword{internal}
