% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_funcs.R
\docType{data}
\name{PUBMED_BASE_ADDRESS}
\alias{PUBMED_BASE_ADDRESS}
\title{PUBMED_BASE_ADDRESS}
\format{
An object of class \code{character} of length 1.
}
\usage{
PUBMED_BASE_ADDRESS
}
\description{
A constant representing the base address for PubMed API queries.
}
\keyword{datasets}
