% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_ncbi_apikey}
\alias{set_ncbi_apikey}
\title{Set NCBI API Key}
\usage{
set_ncbi_apikey(check_env = FALSE)
}
\arguments{
\item{check_env}{Logical. If TRUE, the function will attempt to find the API key in the
"ncbi_apikey" environment variable. If FALSE, the function will prompt the user to enter the API
key interactively. Default is FALSE.}
}
\value{
This function does not return any value. It sets the "ncbi.api.key" session option if the
API key is valid.
}
\description{
This function sets the NCBI API key for use in the current R session. The key can be provided
interactively or retrieved from an environment variable. The validity of the API key is checked
against the NCBI API by making a test request.
}
\examples{
\dontrun{
# Set API key from environment variable
set_ncbi_apikey(check_env = TRUE)

# Set API key interactively
set_ncbi_apikey(check_env = FALSE)
}

}
